requirements.txt
streamlit==1.41.1
chromadb==0.6.3
openai== 1.60.2
tqdm==4.67.1
python-dotenv==1.0.1
langchain==0.3.16
langchain-openai==0.3.2
langchain-experimental==0.3.4
